wip extract methode convertToOccurrenceMap
