+++ result.out
@@ -1,0 +1,23 @@
+tremor = 4
+dizziness = 5
+blindness = 1
+constricted pupils = 3
+fever = 7
+stiff neck = 4
+dilated pupils = 4
+cough = 6
+insomnia = 4
+nausea = 5
+low blood pressure = 4
+rapid heart rate = 1
+dry mouth = 8
+anxiety = 5
+rash = 4
+shortness of breath = 3
+arrhythmias = 3
+inflammation = 7
+high blood pressure = 10
+water retention = 1
+headache = 3
+stomach pain = 3
+blurred vision = 5
