+++ src/main/java/com/hemebiotech/AnalyticsCounter.java
@@ -1,8 +1,8 @@
 package com.hemebiotech;
 
 import com.hemebiotech.fileAccessObject.GetWriteSymptomsOccurrence;
-import com.hemebiotech.model.SymptomReader;
-import com.hemebiotech.model.SymptomReaderImpl;
+import com.hemebiotech.controller.SymptomReader;
+import com.hemebiotech.controller.SymptomReaderImpl;
 
 public class AnalyticsCounter {
     public static void main(String[] args) {
