+++ src/main/java/com/hemebiotech/controller/SymptomReader.java
@@ -1,4 +1,4 @@
-package com.hemebiotech.model;
+package com.hemebiotech.controller;
 
 import java.util.List;
