+++ src/main/java/com/hemebiotech/controller/SymptomReaderImpl.java
@@ -1,4 +1,4 @@
-package com.hemebiotech.model;
+package com.hemebiotech.controller;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
